Major work to make working with IceCore easier
IceCoreGame is now a class you an inherit from instead of the "Game" class you usually do. The IceCoreGame class handles creating the graphics manager, input manager and so on for you.

Inputs are also more versatile now with the "GetActionValue" function. Input are also no longer required to have a callback function.